Working on things

+++ RivenWorldServer/src/main/java/com/gamefocal/rivenworld/commands/chat/admin/LocationCommand.java
@@ -12,7 +12,7 @@ public class LocationCommand extends HiveCommand {
             netConnection.sendChatMessage(ChatColor.BLUE + "Current Location: " + netConnection.getPlayer().location);
             System.out.println("[Loc-cmd]: " + netConnection.getPlayer().location);
 
-            int height = DedicatedServer.instance.getWorld().getRawHeightmap().getHeightFromLocation(netConnection.getPlayer().location);
+            float height = DedicatedServer.instance.getWorld().getRawHeightmap().getHeightFromLocation(netConnection.getPlayer().location);
 
             DedicatedServer.get(RayService.class).makeRequest(netConnection.getPlayer().location, 1, request -> {
                 System.out.println("HM: " + height + ", UE:" + request.getReturnedHeight());
